+++ src-docs/src/views/nav_drawer/nav_drawer.js
@@ -329,6 +329,7 @@ export default class extends Component {
         size: 's',
         style: { color: 'inherit' },
         'aria-label': 'My dashboard',
+        isWrappable: true,
         extraAction: {
           color: 'subdued',
           iconType: 'starEmpty',
@@ -342,7 +343,8 @@ export default class extends Component {
         iconType: 'canvasApp',
         size: 's',
         style: { color: 'inherit' },
-        'aria-label': 'My workpad',
+        'aria-label': 'Workpad with title that wraps',
+        isWrappable: true,
         extraAction: {
           color: 'subdued',
           iconType: 'starEmpty',
@@ -357,6 +359,7 @@ export default class extends Component {
         size: 's',
         style: { color: 'inherit' },
         'aria-label': 'My logs',
+        isWrappable: true,
         extraAction: {
           color: 'subdued',
           iconType: 'starEmpty',